refactor some pieces
